python/requirements.txt
html_sanitizer==1.9.3
bleach==4.1.0
Django==3.2.15
django-html-sanitizer==0.1.5
argparse
flask